Separate test from scalar ALS + fix slight but in for loop not linking to the correct tensor train
